[P045] Don't use UserVar.setInt32() for Collection builds

+++ src/_P045_MPU6050.ino
@@ -271,14 +271,14 @@ boolean Plugin_045(uint8_t function, struct EventStruct *event, String& string)
 
               // Did we count more times exceeded then the minimum detection value?
               if (PCONFIG_LONG(0) >= PCONFIG(5)) {
-                UserVar.setInt32(event->TaskIndex, 0, 1); // x times threshold exceeded within window.
+                UserVar.setFloat(event->TaskIndex, 0, 1.0f); // x times threshold exceeded within window.
               } else {
-                UserVar.setInt32(event->TaskIndex, 0, 0); // reset because x times threshold within window not met.
+                UserVar.setFloat(event->TaskIndex, 0, 0.0f); // reset because x times threshold within window not met.
               }
 
               // Check if UserVar changed so we do not overload homecontroller with the same readings
-              if (PCONFIG(7) != UserVar.getInt32(event->TaskIndex, 0)) {
-                PCONFIG(7) = UserVar.getInt32(event->TaskIndex, 0);
+              if (PCONFIG(7) != UserVar.getFloat(event->TaskIndex, 0)) {
+                PCONFIG(7) = UserVar.getFloat(event->TaskIndex, 0);
                 success    = true;
               } else {
                 success = false;
@@ -297,7 +297,7 @@ boolean Plugin_045(uint8_t function, struct EventStruct *event, String& string)
           {
             uint8_t reqaxis = (_P045_Function - 1) % 3;       // xyz         -> eg: function 5(ay) (5-1) % 3 = 1           (y)
             uint8_t reqvar  = ((_P045_Function - 1) / 3) + 2; // range, a, g -> eg: function 9(gz) ((9-1) / 3 = 2) + 2 = 4 (g)
-            UserVar.setInt32(event->TaskIndex, 0, P045_data->_axis[reqaxis][reqvar]);
+            UserVar.setFloat(event->TaskIndex, 0, P045_data->_axis[reqaxis][reqvar]);
             success                      = true;
             break;
           }
